+++ docs/api/releases.md
@@ -1,12 +1,18 @@
 # Releases
 
-## `list_releases`
+## `list_releases` (GitHub)
+
+Also reachable as `github.list_releases`.
 
 ::: ocx_mirror_sdk.github._router.list_releases
 
 ## `Backend`
 
 ::: ocx_mirror_sdk.github._router.Backend
+
+## `gitlab.list_releases`
+
+::: ocx_mirror_sdk.gitlab._rest.list_releases
 
 ## `Asset`